crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.9.0
pandas==2.1.1
lxml==4.9.3
selenium==4.11.2
pyautogui==0.9.53
alembic==1.11.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
pyyaml==6.0
openai==1.51.0